requirements.txt
google-generativeai
Pillow
opencv-python
python-dotenv
PyYAML
